requirements/base.txt
django==5.2
whitenoise==6.9.0
python-decouple==3.8
djangorestframework==3.16.0
pillow==11.2.1
drf-spectacular==0.28.0
djangorestframework-simplejwt==5.5.0
django-treabeard==4.7.1
drf-nested-routers==0.94.2
adrf==0.1.9
django-cors-headers==4.7.0
locust==2.37.5